Fix infinite loop on OS X when using find in files
The GIOCondition is always set to G_IO_IN even when input end is reached (and no other flags are set) so the read_fif_io() function never returns FALSE which causes an infinite loop. It is necessary to check also the return value of g_io_channel_read_line() and return FALSE on EOF or error.