Fix clang warning regarding mismatching types
The Option initialization didn't match the struct - there was an extra initialization of -e parameter and missing initialization of --etags-include a bit later. This means the initializations of the fields were shifted by one between these two.

I hope this patch won't break anything because of the different initialization (things seem to work).