Remove abc ctags parser
The parser is a joke - I assume it was merged with the hope the author
would work on the commented-out parts which apparently never happened.
In the current form it doesn't do anything useful and by deleting it
we really won't lose any know-how because anyone who'd want to improve it
would have to start from scratch anyway.