Enable the new cxx parser
There are several things needed for this:

1. The new preprocessor has to be defined as a separate parser.

2. Tags from the new c/c++ parsers and the preprocessor parser have to
be mapped to Geany types. We still need to keep the old mappings because
some parsers like Ferite or GLSL still use the old C parser.

3. The cxx parser returns varType inside refType[1] so we have to copy
it from there.

4. Anonymous tags have a different name (in the future we should probably
rename them to something nicer) so we have to reflect this in
tm_tag_is_anon().

For now this patch is meant for experimenting so I haven't updated the
tests some of which fail. However, from the tests I checked the failures
don't seem to be just minor differences in the output rather than real
test fails.