Consistently rename anonymous tags and drop the last ctags diff
Check all the collected tags once a file is parsed and rename them in
the form

anonymousN

where N is the N-th occurrence of an anonymous tag in a file. This way
all anonymous tags will look the same in the symbol tree regardless of
the language so even when subparsers are used, anonymous tags will appear
correctly in the symbol tree and numbering will be preserved.

For instance, in a file with 2 anonymous tags, then 2 anonymous tags in
a subparser and final anonymous tag in the main parser, the tags will look
this way:

anonymous1
anonymous2
anonymous3 - subparser starts here
anonymous4 - subparser ends here
anonymous5

In addition, scopes of the nested tags have to be updated if the parent
tag is an anonymous tag. Also, tags like typedef struct {} Foo; have
to be updated because for the Foo tag, the anonymous struct is reported
as a var_type.