Add option to show symbols in symbol tree without root groups
While the current way of grouping symbols under various categories
like "Functions", "Structures", "Macros" etc. may be useful in some
situations, it doesn't allow true sorting by name or sorting by line
number because this sorting is only valid within the group and not
globally which makes it sometimes hard to find some symbol.

This patch adds the option to switch between two views of the symbol tree
using the context menu:

Categorized - the original view used by Geany
Uncategorized - view with a single root element "Symbols" in which
all symbols are rooted and all of them either sorted by line number
or alphabetically

The core of the implementation is rather simple - instead of using
one of the category roots, NULL is passed as a parent to
gtk_tree_store_insert_with_values() so tags themselves become root
elements. (We still need to keep the category root iters in this situation
to copy the icons for the tags from them but they are not inserted to
the tree).

The rest is some UI stuff for creating entries in the context menu, I also
added a config preference so this settings is remembered across Geany
re-launches. I made this option global for all open files because
I would find it annoying to apply this setting manually for every open
file. So the code also checks whether the view type changed for the
given file when switching tabs and re-creates the tree with the new
setting.

Fixes #1082. (Together with the new Latex parser reporting scope.)